Makefile: relax a bit more the compiler options
I reckon that pict should be compiled with all warnings enabled,
and all warnings being treated as errors.  This is actually the goal
of other changes to be proposed soon.

For now, some code, such as naming attributes with the same name as
previously declared types is not really welcomed by some compilers
such as GCC, being considered "nonconformant code".

Since changes to fix this would bring changes to the pict API, let's
keep things simple and simply allow the unchanged code to build on
other compilers such as GCC.

Signed-off-by: Cleber Rosa <crosa@redhat.com>