Makefile: add source target
It's useful to generate a snapshot of the current state of the GIT
repo.  Let's add a target that generates a tarball using git-archive
using a naming convention that makes it easy to trace to the commit
that generated this snapshot.

This format also happens to be well suitable for using the source
tarball when packaging it in RPM format.

Signed-off-by: Cleber Rosa <crosa@redhat.com>