Makefile: allow parallelization of build
The current rule for building pict is based on a single command
execution, which makes faster builds leveraging parallel execution
impossible.

This change introduces the objects that are required to build pict,
and relies on implicit rules to turn the ".cpp" files into the object
files.

After these changes, a command such as "make -j8" shows the benefits
of this change.

Signed-off-by: Cleber Rosa <crosa@redhat.com>